split out repeted functionality into a base test
